Wordle.Decoder

Description

Documentation

playDecoderMode :: Set Guess -> Set Score -> String -> Int -> IO ()

Ejecuta el modo descifrador del juego.

Argumentos:

El intento de adivinación actual.

El conjunto de posibles adivinaciones en el contexto.

El conjunto de posibles strings de Score del usuario en el contexto.

El número de intentos restantes.

initWordSet :: Set String -> Set Guess

Crea el conjunto inicial con todas las Guess del juego, dadas estas en un conjunto de String.

initScoreSet :: Set Score

Conjunto con todas las combinaciones de Scores posibles.