Wordle.Utils.Checkers

Description

Documentation

Aliases de tipos

type ScoreString = String

String de calificación de una adivinación de Wordle, tiene el siguiente formato:

T donde la letra de la adivinación se encuentra en la posición correspondiente en la respuesta.

V donde la letra de la adivinación se encuentra en la respuesta pero en una posición distinta.

- donde la letra de la adivinación no se encuentra en la respuesta.

type Guess = String

type Answer = String

Interfaz extena del módulo

checkGuess :: Guess -> Answer -> Set String -> Either String ScoreString

Chequea la validez y califica una adivinación de Wordle.

Recibe la adivinación, la respuesta y el Set de palabras válidas y retorna:

Left mensaje si la adivinación es inválida.

Right calificación si la adivinación es válida.

Ejemplos:

>>> checkGuess "CARAS" "CARPA"
Right "TTTV_"
>>> checkGuess "TIENE" "PEROL"
Right "--V--"
>>> checkGuess "PALTA" "RESTO"
Right "---T-"
>>> checkGuess "PERRO" "PERRO"
Right "TTTTT"
>>> checkGuess "TIENE" "TENER"
Right"T-VVV"
>>> checkGuess "SIEMPRE" "NUNCA"
Left "El tamaño de las cadenas no coincide"

Funciones de chequeo

checkT :: Guess -> Answer -> (ScoreString, String)

Chequea los toros de una palabra, dada la adivinación y la respuesta esperada.

Retorna una tupla con:

Una string con las letras restantes por adivinar.

La string de calificación parcial con el formato descrito en checkGuess

checkV :: Guess -> (ScoreString, String) -> ScoreString

Chequea las vacas de la palabra, dado el resultado de haber chequeado los toros con la función checkT previamente.

Retorna una string de calificación.

Funciones de validación

validateGuess :: Guess -> Set String -> Either String Guess

Verifica si una palabra es valida, dadas la palabra y un Set con las palabras válidas.

Retorna:

Left mensaje si la palabra no es válida, con el mensaje de error.

Right palabra si la palabra es válida, con la palabra en mayúsculas y sin acentos.

Una palabra es válida si:

Tiene 5 caracteres.

Todos sus caracteres son alfabéticos

Pertenece al conjunto de palabras válidas.

isAlphabetic :: String -> Bool

Verifica si una String es alfabética.

uniformize :: String -> String

Convierte una String a mayúsculas y remueve los acentos agudos de sus vocales.

Ejemplo:

>>> uniformize "Hola, cómo estás."
"HOLA, COMO ESTAS."